filter._run: Move everything here for now
This is an intermediate step before splitting the file even more.

This module needs to be prefixed with an underscore otherwise there is a
name conflict with the run() function in filter/__init__.py.
