+++ tests/filter/test_run.py
@@ -14,6 +14,7 @@ from Bio.SeqRecord import SeqRecord
 from freezegun import freeze_time
 
 import augur.filter
+import augur.filter._run
 from augur.io.metadata import read_metadata
 from augur.errors import AugurError
 
@@ -68,7 +69,7 @@ def mock_priorities_file_valid_with_spaces_and_tabs(mocker):
 class TestFilter:
     def test_read_priority_scores_valid(self, mock_priorities_file_valid):
         # builtins.open is stubbed, but we need a valid file to satisfy the existence check
-        priorities = augur.filter.read_priority_scores(
+        priorities = augur.filter._run.read_priority_scores(
             "tests/builds/tb/data/lee_2015.vcf"
         )
 
@@ -79,11 +80,11 @@ class TestFilter:
     def test_read_priority_scores_malformed(self, mock_priorities_file_malformed):
         with pytest.raises(ValueError):
             # builtins.open is stubbed, but we need a valid file to satisfy the existence check
-            augur.filter.read_priority_scores("tests/builds/tb/data/lee_2015.vcf")
+            augur.filter._run.read_priority_scores("tests/builds/tb/data/lee_2015.vcf")
 
     def test_read_priority_scores_valid_with_spaces_and_tabs(self, mock_priorities_file_valid_with_spaces_and_tabs):
         # builtins.open is stubbed, but we need a valid file to satisfy the existence check
-        priorities = augur.filter.read_priority_scores(
+        priorities = augur.filter._run.read_priority_scores(
             "tests/builds/tb/data/lee_2015.vcf"
         )
 
@@ -91,7 +92,7 @@ class TestFilter:
 
     def test_read_priority_scores_does_not_exist(self):
         with pytest.raises(FileNotFoundError):
-            augur.filter.read_priority_scores("/does/not/exist.txt")
+            augur.filter._run.read_priority_scores("/does/not/exist.txt")
 
     def test_filter_on_query_good(self, tmpdir, sequences):
         """Basic filter_on_query test"""
@@ -100,7 +101,7 @@ class TestFilter:
                                           ("SEQ_2","colorado","bad"),
                                           ("SEQ_3","nevada","good")))
         metadata = read_metadata(meta_fn)
-        filtered = augur.filter.filter_by_query(metadata, 'quality=="good"')
+        filtered = augur.filter._run.filter_by_query(metadata, 'quality=="good"')
         assert sorted(filtered) == ["SEQ_1", "SEQ_3"]
 
     def test_filter_run_with_query(self, tmpdir, fasta_fn, argparser):
@@ -112,7 +113,7 @@ class TestFilter:
                                           ("SEQ_3","nevada","good")))
         args = argparser('-s %s --metadata %s -o %s --query "location==\'colorado\'"'
                          % (fasta_fn, meta_fn, out_fn))
-        augur.filter.run(args)
+        augur.filter._run.run(args)
         output = SeqIO.to_dict(SeqIO.parse(out_fn, "fasta"))
         assert list(output.keys()) == ["SEQ_1", "SEQ_2"]
 
@@ -127,7 +128,7 @@ class TestFilter:
         open(include_fn, "w").write("SEQ_3")
         args = argparser('-s %s --metadata %s -o %s --query "quality==\'good\' & location==\'colorado\'" --include %s'
                          % (fasta_fn, meta_fn, out_fn, include_fn))
-        augur.filter.run(args)
+        augur.filter._run.run(args)
         output = SeqIO.to_dict(SeqIO.parse(out_fn, "fasta"))
         assert list(output.keys()) == ["SEQ_1", "SEQ_3"]
 
@@ -140,7 +141,7 @@ class TestFilter:
                                           ("SEQ_3","nevada","good")))
         args = argparser('-s %s --metadata %s -o %s --query "quality==\'good\' & location==\'colorado\'" --include-where "location=nevada"'
                          % (fasta_fn, meta_fn, out_fn))
-        augur.filter.run(args)
+        augur.filter._run.run(args)
         output = SeqIO.to_dict(SeqIO.parse(out_fn, "fasta"))
         assert list(output.keys()) == ["SEQ_1", "SEQ_3"]
 
@@ -154,7 +155,7 @@ class TestFilter:
                                           ("SEQ_3","2020-02-25")))
         args = argparser('-s %s --metadata %s -o %s --min-date %s'
                          % (fasta_fn, meta_fn, out_fn, min_date))
-        augur.filter.run(args)
+        augur.filter._run.run(args)
         output = SeqIO.to_dict(SeqIO.parse(out_fn, "fasta"))
         assert list(output.keys()) == ["SEQ_1", "SEQ_2"]
 
@@ -168,7 +169,7 @@ class TestFilter:
                                           ("SEQ_3","2020-03-02")))
         args = argparser('-s %s --metadata %s -o %s --max-date %s'
                          % (fasta_fn, meta_fn, out_fn, max_date))
-        augur.filter.run(args)
+        augur.filter._run.run(args)
         output = SeqIO.to_dict(SeqIO.parse(out_fn, "fasta"))
         assert list(output.keys()) == ["SEQ_1", "SEQ_2"]
 
@@ -182,7 +183,7 @@ class TestFilter:
                                           ("SEQ_3","2020-XX-XX")))
         args = argparser('-s %s --metadata %s -o %s --min-date %s'
                          % (fasta_fn, meta_fn, out_fn, min_date))
-        augur.filter.run(args)
+        augur.filter._run.run(args)
         output = SeqIO.to_dict(SeqIO.parse(out_fn, "fasta"))
         assert list(output.keys()) == ["SEQ_2", "SEQ_3"]
 
@@ -196,7 +197,7 @@ class TestFilter:
                                           ("SEQ_3","2020-XX-XX")))
         args = argparser('-s %s --metadata %s -o %s --min-date %s'
                          % (fasta_fn, meta_fn, out_fn, min_date))
-        augur.filter.run(args)
+        augur.filter._run.run(args)
         output = SeqIO.to_dict(SeqIO.parse(out_fn, "fasta"))
         assert list(output.keys()) == ["SEQ_1", "SEQ_2", "SEQ_3"]
 
@@ -320,7 +321,7 @@ class TestFilter:
         meta_fn = write_metadata(tmpdir, (("strain","date"),
                                           *metadata_rows))
         args = argparser(f'--metadata {meta_fn} --output-strains {out_fn} {argparse_params}')
-        augur.filter.run(args)
+        augur.filter._run.run(args)
         with open(out_fn) as f:
             output_sorted = sorted(line.rstrip() for line in f)
         assert output_sorted == output_sorted_expected
@@ -366,7 +367,7 @@ class TestFilterGroupBy:
     def test_filter_groupby_strain_subset(self, valid_metadata: pd.DataFrame):
         metadata = valid_metadata.copy()
         strains = ['SEQ_1', 'SEQ_3', 'SEQ_5']
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata)
         assert group_by_strain == {
             'SEQ_1': ('_dummy',),
             'SEQ_3': ('_dummy',),
@@ -377,7 +378,7 @@ class TestFilterGroupBy:
     def test_filter_groupby_dummy(self, valid_metadata: pd.DataFrame):
         metadata = valid_metadata.copy()
         strains = metadata.index.tolist()
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata)
         assert group_by_strain == {
             'SEQ_1': ('_dummy',),
             'SEQ_2': ('_dummy',),
@@ -392,14 +393,14 @@ class TestFilterGroupBy:
         metadata = valid_metadata.copy()
         strains = metadata.index.tolist()
         with pytest.raises(AugurError) as e_info:
-            augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+            augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert str(e_info.value) == "The specified group-by categories (['invalid']) were not found."
 
     def test_filter_groupby_invalid_warn(self, valid_metadata: pd.DataFrame, capsys):
         groups = ['country', 'year', 'month', 'invalid']
         metadata = valid_metadata.copy()
         strains = metadata.index.tolist()
-        group_by_strain, _ = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, _ = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {
             'SEQ_1': ('A', 2020, (2020, 1), 'unknown'),
             'SEQ_2': ('A', 2020, (2020, 2), 'unknown'),
@@ -415,7 +416,7 @@ class TestFilterGroupBy:
         metadata = valid_metadata.copy()
         metadata.at["SEQ_2", "date"] = "XXXX-02-01"
         strains = metadata.index.tolist()
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {
             'SEQ_1': ('A', 2020, (2020, 1)),
             'SEQ_3': ('B', 2020, (2020, 3)),
@@ -429,7 +430,7 @@ class TestFilterGroupBy:
         metadata = valid_metadata.copy()
         metadata.at["SEQ_2", "date"] = None
         strains = metadata.index.tolist()
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {
             'SEQ_1': ('A', 2020, (2020, 1)),
             'SEQ_3': ('B', 2020, (2020, 3)),
@@ -443,7 +444,7 @@ class TestFilterGroupBy:
         metadata = valid_metadata.copy()
         metadata.at["SEQ_2", "date"] = "2020-XX-01"
         strains = metadata.index.tolist()
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {
             'SEQ_1': ('A', 2020, (2020, 1)),
             'SEQ_3': ('B', 2020, (2020, 3)),
@@ -457,7 +458,7 @@ class TestFilterGroupBy:
         metadata = valid_metadata.copy()
         metadata.at["SEQ_2", "date"] = "2020"
         strains = metadata.index.tolist()
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {
             'SEQ_1': ('A', 2020, (2020, 1)),
             'SEQ_3': ('B', 2020, (2020, 3)),
@@ -472,7 +473,7 @@ class TestFilterGroupBy:
         metadata = metadata.drop('date', axis='columns')
         strains = metadata.index.tolist()
         with pytest.raises(AugurError) as e_info:
-            augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+            augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert str(e_info.value) == "The specified group-by categories (['year']) were not found. Note that using any of ['month', 'week', 'year'] requires a column called 'date'."
 
     def test_filter_groupby_missing_month_error(self, valid_metadata: pd.DataFrame):
@@ -481,7 +482,7 @@ class TestFilterGroupBy:
         metadata = metadata.drop('date', axis='columns')
         strains = metadata.index.tolist()
         with pytest.raises(AugurError) as e_info:
-            augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+            augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert str(e_info.value) == "The specified group-by categories (['month']) were not found. Note that using any of ['month', 'week', 'year'] requires a column called 'date'."
 
     def test_filter_groupby_missing_year_and_month_error(self, valid_metadata: pd.DataFrame):
@@ -490,7 +491,7 @@ class TestFilterGroupBy:
         metadata = metadata.drop('date', axis='columns')
         strains = metadata.index.tolist()
         with pytest.raises(AugurError) as e_info:
-            augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+            augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert str(e_info.value) == "The specified group-by categories (['year', 'month']) were not found. Note that using any of ['month', 'week', 'year'] requires a column called 'date'."
 
     def test_filter_groupby_missing_date_warn(self, valid_metadata: pd.DataFrame, capsys):
@@ -498,7 +499,7 @@ class TestFilterGroupBy:
         metadata = valid_metadata.copy()
         metadata = metadata.drop('date', axis='columns')
         strains = metadata.index.tolist()
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {
             'SEQ_1': ('A', 'unknown', 'unknown'),
             'SEQ_2': ('A', 'unknown', 'unknown'),
@@ -514,7 +515,7 @@ class TestFilterGroupBy:
         groups = ['country', 'year', 'month']
         metadata = valid_metadata.copy()
         strains = []
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {}
         assert skipped_strains == []
 
@@ -523,7 +524,7 @@ class TestFilterGroupBy:
         metadata = valid_metadata.copy()
         metadata['date'] = '2020'
         strains = metadata.index.tolist()
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {
             'SEQ_1': ('A', 2020),
             'SEQ_2': ('A', 2020),
@@ -538,7 +539,7 @@ class TestFilterGroupBy:
         metadata = valid_metadata.copy()
         metadata['date'] = '2020'
         strains = metadata.index.tolist()
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {}
         assert skipped_strains == [
             {'strain': 'SEQ_1', 'filter': 'skip_group_by_with_ambiguous_month', 'kwargs': ''},
@@ -553,7 +554,7 @@ class TestFilterGroupBy:
         metadata = valid_metadata.copy()
         metadata['date'] = '2020-01'
         strains = metadata.index.tolist()
-        group_by_strain, skipped_strains = augur.filter.get_groups_for_subsampling(strains, metadata, group_by=groups)
+        group_by_strain, skipped_strains = augur.filter._run.get_groups_for_subsampling(strains, metadata, group_by=groups)
         assert group_by_strain == {
             'SEQ_1': ('A', 2020, (2020, 1)),
             'SEQ_2': ('A', 2020, (2020, 1)),
